requirements.txt
requests>=2.31.0
pandas>=2.0.0
python-dotenv>=1.0.0
streamlit>=1.31.0
Pillow
matplotlib
openai>=1.12.0
plotly>=5.18.0
seaborn
altair>=5.2.0
